I have returned
Just some ideas.

Refamiliarizing myself with the project and have decided to try and flesh out the Skill-landscape feature again.
Added files are volatile as I decide whether I want an external manager type object managing all of the player's progress, or another component on the player.

Probably going to collapse this branch and open a new one tomorrow.