Hex and skills progression push
Ive spent a lot of time upgrading the ascetic of the UI, but I have been struggling with finding a visually pleasing way to display the current state of the Hex grid skill map. For now, I have decided that the Hex Grid will only be accessible while asleep, in which case the player controller will unpossess the current character, and possess a separate pawn that will give a camera view and controls over the Hex Grid progression system.

I have added some dummy DND based skill trees to populate in the future for prototyping the Hex Grid system.
I have fleshed out the Progression Manager Component to spawn in and provide access to a Progression Manager Actor. Intended to be added to the Player Controller.
This Progression Manager should have access to the Hex Grid Progression system, all acquired Skill Trees, and any future progression systems such as a looting system.

My next upcoming push should be a Documentation Commit, in which I will move all current Game Design Documentation from Notion, to this project as a dedicated local congregation of files.