Skill Tree Generic Component Drop Down Menu Success!!!
Generic Component Drop Down now includes Blueprint and C++ component classes!

Now we are focusing on adding the selected component instance to the Node, and displaying the current components on the node as well.

This would be an equivalent to the "Simple Construction Script," and "Subobject Editor" objects/systems that Blueprints use for the Actor/Scene Component system.